Feat: #61 CommunityPost 라우트에 type과 ID 파라미터 추가
