+++ src/App.tsx
@@ -10,7 +10,6 @@ import MypageEditor from './pages/mypage/MyPageEditor';
 import LogIn from './pages/login/LogIn';
 import SignUp from './pages/signup/SignUp';
 import CommunityPost from './pages/community/CommunityPost';
-import CommunityEdit from './pages/community/CommunityEdit';
 import Layout from './layout/Layout';
 
 const App = () => {
@@ -20,8 +19,14 @@ const App = () => {
         <Route path="/" element={<Main />} />
         <Route path="/community" element={<Community />} />
         <Route path="/communitydetail" element={<CommunityDetail />} />
-        <Route path="/communityedit" element={<CommunityEdit />} />
-        <Route path="/communitypost" element={<CommunityPost />} />
+        <Route
+          path="/communitypost:id"
+          element={<CommunityPost type={'edit'} />}
+        />
+        <Route
+          path="/communitypost"
+          element={<CommunityPost type={'post'} />}
+        />
         <Route path="/contact" element={<Contact />} />
         <Route path="/contactform" element={<ContactForm />} />
         <Route path="/mypage" element={<Mypage />} />
